backend/src/requirements.txt
pymongo==3.11.2
selenium==3.141.0
beautifulsoup4==4.9.3
psycopg2==2.8.6
python-dotenv==0.15.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
lxml==4.6.2
requests==2.25.1
soupsieve==2.1
urllib3==1.26.3
webdriver-manager==3.3.0
